Additional room reservation checks

See merge request cse1105/2019-2020/repos/oopp-group-62/repository-template!65

+++ server/src/main/java/nl/tudelft/oopp/demo/controllers/RoomReservationController.java
@@ -3,6 +3,7 @@ package nl.tudelft.oopp.demo.controllers;
 import java.time.LocalDate;
 import java.time.LocalTime;
 import java.time.format.DateTimeFormatter;
+import java.time.temporal.WeekFields;
 import java.util.ArrayList;
 import java.util.HashMap;
 import java.util.List;
@@ -329,6 +330,15 @@ public class RoomReservationController {
                 return true;
             }
         }
+
+        // if the user has reserved this room at some point during the week
+        int weekOfNewReservation = newRoomReservation.getDate().get(WeekFields.ISO.weekOfWeekBasedYear());
+        for (RoomReservation reservation : reservations.findByUserIdAndRoomId(newRoomReservation.getUser().getId(), newRoomReservation.getRoom().getId())) {
+            int weekOfRes = reservation.getDate().plusDays(1).get(WeekFields.ISO.weekOfWeekBasedYear());
+            if (weekOfNewReservation == weekOfRes && reservation.getId() != newRoomReservation.getId() && reservation.getDate().getYear() == newRoomReservation.getDate().getYear()) {
+                return true;
+            }
+        }
         return false;
     }
 
@@ -363,6 +373,13 @@ public class RoomReservationController {
                     return new ResponseEntity<RoomReservation>(HttpStatus.CONFLICT);
                 }
 
+                // users should not be able to delete room reservations for previous days
+                if (newRoomReservation.getDate().compareTo(LocalDate.now()) < 0
+                    || (LocalDate.now().equals(newRoomReservation.getDate())
+                    && newRoomReservation.getStartTime().compareTo(LocalTime.now()) < 0)) {
+                    return new ResponseEntity<RoomReservation>(HttpStatus.CONFLICT);
+                }
+
                 roomReservation.setUser(newRoomReservation.getUser());
                 roomReservation.setRoom(newRoomReservation.getRoom());
                 roomReservation.setDate(newRoomReservation.getDate());
@@ -393,6 +410,13 @@ public class RoomReservationController {
 
         if (!reservationToDelete.getUser().getUsername().equals(authentication.getName())) {
             return new ResponseEntity<>(HttpStatus.UNAUTHORIZED);
+        }
+
+        // users should not be able to delete room reservations for previous days
+        if (reservationToDelete.getDate().plusDays(1).compareTo(LocalDate.now()) < 0
+            || (LocalDate.now().equals(reservationToDelete.getDate().plusDays(1))
+            && reservationToDelete.getStartTime().compareTo(LocalTime.now()) < 0)) {
+            return new ResponseEntity<>(HttpStatus.CONFLICT);
         }
 
         reservations.deleteById(roomReservationId);
